# Fix performance issue in dask.graph_manipulation.wait_on()