wait_on() is affected by a severe performance degradation issue caused by an incorrect usage of blockwise().

To reproduce:

```bash
dask-worker localhost:8786 --nthreads 1 --nprocs 150 --memory-limit 370MiB
```
(you'll need 64 GiB RAM)

```python
import dask.array as da
import dask.graph_manipulation as gm
from distributed import Client

client = Client("localhost:8786")
%time gm.wait_on(da.random.random(150, chunks=1)).sum().compute()
%time gm.wait_on(da.random.random(150, chunks=1).persist()).sum().compute()
```
|  | master | this PR |
| :-- | --: | --: |
| without persist() | 1.62s | 587 ms |
| with persist() | 12.3s | 604 ms |