@jrbourbeau 2021-02-23
These methods haven't been released yet, but will be included in the next release (hopefully this Friday)